Changed big decimal to double for amobee tracking

+++ project/src/com/expedia/bookings/activity/FlightBookingActivity.java
@@ -288,7 +288,7 @@ public class FlightBookingActivity extends SherlockFragmentActivity implements C
 						}
 						Money money = Db.getFlightSearch().getSelectedFlightTrip().getTotalFare();
 						if (money != null) {
-							Amobee.trackCheckout(money.getCurrency(), money.getAmount(), days);
+							Amobee.trackCheckout(money.getCurrency(), money.getAmount().doubleValue(), days);
 						}
 					}
 				}
